image_project/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
opencv-contrib-python>=4.8.0
scikit-learn>=1.5.0
numpy>=1.26.0
joblib>=1.3.0
google-genai
python-dotenv>=1.0.0
Pillow>=10.0.0
